Move Version2CustomTypeAdapterToAdapter to internal package, and add a requires optin `ApolloInternal` annotation on it
