+++ apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/ApolloInternal.kt
@@ -1,0 +1,9 @@
+package com.apollographql.apollo3.api
+
+@RequiresOptIn(
+    level = RequiresOptIn.Level.ERROR,
+    message = "This API is for internal use only in Apollo"
+)
+@Retention(AnnotationRetention.BINARY)
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
+annotation class ApolloInternal
